requirements.txt
asgiref==3.3.1
attrs==20.3.0
AudioConverter==0.3.0
Automat==20.2.0
Babel==2.9.0
beautifulsoup4==4.9.3
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
constantly==15.1.0
Django==3.1.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-extensions==3.1.0
django-ipware==3.0.2
dlib==19.21.99
dnspython==1.16.0
docopt==0.6.2
face-recognition==1.3.0
face-recognition-models==0.3.0
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
ipython-genutils==0.2.0
jsonschema==3.2.0
jupyter-core==4.7.0
nbformat==5.1.2
NetfilterQueue==0.8.1
nmap==0.0.1
numpy==1.19.5
oauthlib==3.1.0
pigar==0.10.0
Pillow==8.1.0
pipreqs==0.4.10
pocketsphinx==0.1.15
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.24.1
PyHamcrest==2.0.2
pymongo==3.11.2
pyrsistent==0.17.3
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scapy==2.4.4
six==1.15.0
soupsieve==2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
tkcalendar==1.6.1
traitlets==5.0.5
Twisted==20.3.0
uritemplate==3.0.1
urllib3==1.26.2
vosk==0.3.17
yarg==0.1.9
zope.interface==5.2.0
